requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Caching==2.3.0
Flask-Limiter==3.7.0
python-dotenv==1.0.1
razorpay==1.5.0
reportlab==4.2.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.7
SQLAlchemy==2.0.30
alembic==1.13.1
WTForms==3.1.2
email-validator==2.1.1
requests==2.32.3
cryptography==42.0.8
psutil==5.9.8
Chart.js==4.4.3